refactor: Improve reservation creation logic and structure optimize: Separate concerns for better reservation component handling
